Rewriite entire source (limine -> multiboot2)
As part of the move to the multiboot2 protocol, the project source has been restructured and refactored. Limine for SOME reason doesn't map the RSDP into the pagemap, which means reading the RSDP and any subsequent ACPI data causes page faults and creates headaches. We've moved to Grub. Huge thanks to @LoxyMono for helping to port the kernel to multiboot2.